refactor(painters/dom): colocate paragraph borders and marker helpers
Move `features/paragraph-borders/` to `paragraph/borders/` and fold
`utils/marker-helpers.ts` into `paragraph/list-marker.ts` so all
paragraph rendering pieces live under `paragraph/`. Update the
feature registry and imports to match.
