+++ packages/layout-engine/painters/dom/src/features/feature-registry.ts
@@ -17,8 +17,8 @@ export const RENDERING_FEATURES = {
   // ─── Paragraph Borders ───────────────────────────────────────────
   // @spec ECMA-376 §17.3.1.24 (pBdr)
   'w:pBdr': {
-    feature: 'paragraph-borders',
-    module: './paragraph-borders',
+    feature: 'paragraph/borders',
+    module: '../paragraph/borders',
     handles: ['w:pBdr/w:top', 'w:pBdr/w:bottom', 'w:pBdr/w:left', 'w:pBdr/w:right', 'w:pBdr/w:between', 'w:pBdr/w:bar'],
     spec: '§17.3.1.24',
   },
@@ -26,8 +26,8 @@ export const RENDERING_FEATURES = {
   // ─── Paragraph Shading ───────────────────────────────────────────
   // @spec ECMA-376 §17.3.1.31 (shd)
   'w:shd': {
-    feature: 'paragraph-borders', // shading shares the border layer module
-    module: './paragraph-borders',
+    feature: 'paragraph/borders', // shading shares the border layer module
+    module: '../paragraph/borders',
     handles: ['w:shd/@w:fill', 'w:shd/@w:val', 'w:shd/@w:color'],
     spec: '§17.3.1.31',
   },
