+++ packages/layout-engine/painters/dom/src/paragraph/borders/border-layer.ts
