+++ packages/layout-engine/painters/dom/src/paragraph/borders/group-analysis.ts
