+++ packages/layout-engine/painters/dom/src/paragraph/borders/index.ts
